added add course step 2

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section=require("../models/Section");
 const Course=require("../models/Course");
+const SubSection = require("../models/SubSection");
 
 
 exports.createSection=async (req, res)=>{
@@ -23,13 +24,19 @@ exports.createSection=async (req, res)=>{
                                                         courseContent:newSection._id,
                                                     }
                                                 },
-                                                {new:true},//HW: use populatae to replace sectin and subsection both in the updated coutse
-        )
+                                                {new:true},
+                                            ).populate({
+                                                path: "courseContent",
+                                                populate: {
+                                                    path: "subSection",
+                                                },
+                                            })
+                                            .exec();
         //return response
         return res.status(200).json({
             success:true,
             message:"Section Succesfully Created",
-            updatedCourseDetails,
+            data:updatedCourseDetails,
         })
 
 
@@ -47,7 +54,7 @@ exports.updateSection= async (req,res)=>{
     try{
         //Data Inputs
         console.log("CheckPOint");
-        const {sectionName,sectionId}=req.body;
+        const {sectionName,sectionId,courseId}=req.body;
 
         //Data Validations
         if(!sectionName || !sectionId){
@@ -57,14 +64,23 @@ exports.updateSection= async (req,res)=>{
             })
         }
         //UPdate Data
-        const section=await Section.findByIdAndUpdate(sectionId,{sectionName},{new:true})
+        const section=await Section.findByIdAndUpdate(
+                                        sectionId,
+                                        {sectionName},{new:true})
         console.log("CheckPOint1");
+        const course=await Course.findById(courseId)
+        .populate({
+            path:"courseContent",
+            populate:{
+                path:"subSection"
+            }
+        }).exec()
 
         //return response
         return res.status(200).json({
             success:true,
-            message:"Section Updated Succesfully",
-            section,
+            message:section,
+            data:course,
         })
 
 
@@ -81,21 +97,47 @@ exports.updateSection= async (req,res)=>{
 
 exports.deleteSection=async (req,res)=>{
     try{
-        const {sectionId}=req.body;
-        console.log(sectionId)
-        await Section.findByIdAndDelete(sectionId);
-        //todo do we need to delete the entry from the course schema  "will do in testing"
-        console.log("checkpoint")
-        return res.status(200).json({
-            success:true,
-            message:"Section Deleted Succesfully"
-        })
+        const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
+
+		await Section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
+		res.status(200).json({
+			success:true,
+			message:"Section deleted",
+			data:course
+		});
+
     }
     catch(error){
-        return res.status(400).json({
-            success:false,
-            message:"Something Went Wrong"
-        })
-
+        console.error("Error deleting section:", error);
+		res.status(500).json({
+			success: false,
+			message: "Internal server error",
+		});
     }
 }
